finished the algorithm to find the winners and the losers and some front end changes

+++ API/Repos/Dtos/CreateEasyDrawDto.cs
@@ -6,7 +6,7 @@
         public DateTime? StartOn { get; set; } 
         public DateTime? EndOn { get; set; } 
         public ulong? CustStatus { get; set; }
-        public int? TicketNo { get; set; }
+        public long? TicketNo { get; set; }
         public string? RaffleName { get; set; }
         public AuthDto AuthDto { get; set; }
     }
